build: migrate from Next.js to Vite
Replace Next.js 16 + App Router with Vite 6 + react-router-dom v7 in
SPA mode. Tailwind v3 → v4 (CSS-side config). Drop sass in
kprfordummies in favour of native CSS nesting + custom properties.

* bilbatez.dev: index.html + src/main.tsx entry, app/ holds plain
  React components, ExternalRedirect routes replace next.config
  redirects for /github /linkedin /bofa /shopee /blibli.
* kprfordummies: same structure; globals.scss → globals.css;
  ExternalRedirect for /bilbatez.dev.
* ESLint flat config: drop @next/eslint-plugin-next, add
  eslint-plugin-react + eslint-plugin-react-hooks; disable
  rules-of-hooks for Playwright tests/.
* Playwright pinned to 1.56.1 — newer Chromium V8/ICU drops the
  ",00" cents from Intl.NumberFormat('id-ID', { currency: 'IDR' }),
  breaking currency assertions.
* CLAUDE.md: refreshed stack notes and added explicit Conventional
  Commits requirement for Claude Code contributions.
