+++ .github/workflows/lint.yml
@@ -25,6 +25,8 @@ jobs:
         run: |
           bun install conventional-changelog-conventionalcommits
           bun install commitlint@latest
+        env:
+          PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: 1
 
       - name: Validate current commit (last commit) with commitlint
         if: github.event_name == 'push'
@@ -50,5 +52,7 @@ jobs:
           bun-version: latest
       - name: Install dependencies
         run: bun install
+        env:
+          PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: 1
       - name: Run linter
         run: bun eslint
