+++ bilbatez.dev/app/NotFound.tsx
@@ -1,3 +1,3 @@
-export default function NotFound() {
+export function NotFound() {
   return <>Oops... seems like there&apos;s nothing here!</>;
 }
