+++ bilbatez.dev/public/manifest.webmanifest
@@ -1,0 +1,8 @@
+{
+  "name": "bilbatez.dev",
+  "short_name": "bilbatez.dev",
+  "description": "Personal Website | Bilbatez.dev",
+  "start_url": "/",
+  "background_color": "#fff",
+  "theme_color": "#fff"
+}
