--- bilbatez.dev/tailwind.config.ts
@@ -1,21 +1,0 @@
-import type { Config } from 'tailwindcss';
-
-import defaultTheme from 'tailwindcss/defaultTheme';
-
-const config: Config = {
-  content: [
-    './pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './components/**/*.{js,ts,jsx,tsx,mdx}',
-    './app/**/*.{js,ts,jsx,tsx,mdx}',
-  ],
-  theme: {
-    extend: {
-      fontFamily: {
-        serif: ['MesloLGS-NF', ...defaultTheme.fontFamily.serif],
-        sans: [...defaultTheme.fontFamily.sans],
-      },
-    },
-  },
-  plugins: [],
-};
-export default config;
