+++ bilbatez.dev/vite.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import tailwindcss from '@tailwindcss/vite';
+
+export default defineConfig({
+  plugins: [react(), tailwindcss()],
+  server: {
+    host: '127.0.0.1',
+    port: 3001,
+    strictPort: true,
+  },
+  preview: {
+    host: '127.0.0.1',
+    port: 3001,
+    strictPort: true,
+  },
+});
