+++ kprfordummies/app/ExternalRedirect.tsx
@@ -1,0 +1,8 @@
+import { useEffect } from 'react';
+
+export function ExternalRedirect({ to }: { to: string }) {
+  useEffect(() => {
+    window.location.replace(to);
+  }, [to]);
+  return null;
+}
