+++ kprfordummies/app/components/Footer.tsx
@@ -1,11 +1,16 @@
-import Link from 'next/link';
 import { memo } from 'react';
+import { Link } from 'react-router-dom';
 
 function Footer() {
   return (
     <footer className="mt-5">
       Dibuat oleh
-      <Link href={'bilbatez.dev'} target="_blank" className="footer-hero">
+      <Link
+        to="/bilbatez.dev"
+        target="_blank"
+        rel="noopener noreferrer"
+        className="footer-hero"
+      >
         Bilbatez.dev 👽
       </Link>
     </footer>
