--- kprfordummies/next.config.mjs
@@ -1,14 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  async redirects() {
-    return [
-      {
-        source: '/bilbatez.dev',
-        destination: 'https://bilbatez.dev',
-        permanent: false,
-      },
-    ];
-  },
-};
-
-export default nextConfig;
