+++ kprfordummies/package.json
@@ -3,8 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev -p 3002",
-    "prod": "next build && next start -p 3002",
+    "dev": "vite --port 3002 --strictPort",
+    "build": "tsc -b && vite build",
+    "preview": "vite preview --port 3002 --strictPort",
+    "prod": "bun run build && bun run preview",
     "test": "cd .. && bun playwright test tests/kprfordummies"
   },
   "dependencies": {
