--- kprfordummies/postcss.config.mjs
@@ -1,10 +1,0 @@
-/** @type {import('postcss-load-config').Config} */
-const config = {
-  plugins: {
-    'postcss-import': {},
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
-
-export default config;
