+++ kprfordummies/vite.config.ts
@@ -1,0 +1,23 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import tailwindcss from '@tailwindcss/vite';
+import path from 'node:path';
+
+export default defineConfig({
+  plugins: [react(), tailwindcss()],
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, '.'),
+    },
+  },
+  server: {
+    host: '127.0.0.1',
+    port: 3002,
+    strictPort: true,
+  },
+  preview: {
+    host: '127.0.0.1',
+    port: 3002,
+    strictPort: true,
+  },
+});
