+++ package.json
@@ -7,43 +7,42 @@
     "./kprfordummies"
   ],
   "scripts": {
-    "postinstall": "bun playwright install",
+    "postinstall": "[ \"$PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD\" = \"1\" ] || bun playwright install",
     "test": "bun playwright test",
     "test:ui": "bun playwright test --ui",
-    "clean": "find . -type d \\( -name 'node_modules' -o -name '.next' \\) -exec rm -rf {} +",
+    "clean": "find . -type d \\( -name 'node_modules' -o -name 'dist' \\) -exec rm -rf {} +",
     "refresh": "bun clean && bun install",
     "prepare": "husky",
     "format": "prettier --write ."
   },
   "dependencies": {
-    "next": "^16.0.7",
     "react": "^19.2.1",
-    "react-dom": "^19.2.1"
+    "react-dom": "^19.2.1",
+    "react-router-dom": "^7.1.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^19.8.1",
-    "@commitlint/config-conventional": "^19.8.1",
-    "@eslint/js": "^9.38.0",
-    "@next/eslint-plugin-next": "^16.0.0",
-    "@playwright/test": "^1.56.1",
-    "@types/node": "^20.19.23",
-    "@types/react": "^19.2.2",
-    "@types/react-dom": "^19.2.2",
-    "autoprefixer": "^10.4.21",
-    "eslint": "^9.38.0",
-    "eslint-config-next": "^16.0.0",
+    "@commitlint/cli": "^21.0.1",
+    "@commitlint/config-conventional": "^21.0.1",
+    "@eslint/js": "^9.39.4",
+    "@playwright/test": "1.56.1",
+    "@tailwindcss/vite": "^4.0.0",
+    "@types/node": "^20.19.41",
+    "@types/react": "^19.2.15",
+    "@types/react-dom": "^19.2.3",
+    "@vitejs/plugin-react": "^4.7.0",
+    "eslint": "^9.39.4",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-playwright": "^2.2.2",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "husky": "^9.1.7",
     "lint-staged": "^15.5.2",
-    "postcss": "^8.5.6",
-    "postcss-import": "^16.1.1",
-    "prettier": "3.5.2",
-    "sass": "^1.93.2",
-    "tailwindcss": "^3.4.18",
+    "prettier": "3.8.3",
+    "tailwindcss": "^4.0.0",
     "tsconfig-paths": "^4.2.0",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.46.2"
+    "typescript-eslint": "^8.60.0",
+    "vite": "^6.4.2"
   },
   "lint-staged": {
     "**/*": [
@@ -52,7 +51,7 @@
     ]
   },
   "overrides": {
-    "@types/react": "19.2.2",
-    "@types/react-dom": "19.2.2"
+    "@types/react": "19.2.15",
+    "@types/react-dom": "19.2.3"
   }
 }
